+++ src/fps_meter.cpp
@@ -25,7 +25,7 @@ void FPSMeter::frameEnd()
     m_framesReallyDone = framesDone;
     const uint64_t usec = std::chrono::duration_cast<std::chrono::microseconds>( m_accumulatedDuration ).count();
     m_accumulatedDuration = std::chrono::seconds( 0 );
-    const float averageFrameDuration = static_cast<float>( usec ) / framesDone;
+    const float averageFrameDuration = static_cast<float>( usec ) / static_cast<float>( framesDone );
     m_lastFpsValue = 1'000'000.0f / averageFrameDuration;
 }
